requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
altex==0.2.0
annotated-types==0.7.0
anyio==4.11.0
asn1crypto==1.5.1
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
boto3==1.40.75
botocore==1.40.75
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
defusedxml==0.7.1
entrypoints==0.4
et_xmlfile==2.0.0
Faker==38.0.0
favicon==0.7.0
feature_engine==1.9.3
filelock==3.20.0
filetype==1.2.0
fonttools==4.60.1
fpdf2==2.8.5
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.6
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-genai==1.51.0
google-generativeai==0.7.2
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
htbuilder==0.9.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==0.3.0
langchain-core==0.3.63
langchain-google-genai==2.0.0
langchain-text-splitters==0.3.8
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.1.147
lxml==6.0.2
Markdown==3.10
markdownlit==0.0.7
MarkupSafe==3.0.3
matplotlib==3.10.7
memory==1.0.0
multidict==6.7.0
narwhals==2.12.0
numpy==1.26.4
openpyxl==3.1.5
orjson==3.11.4
ormsgpack==1.12.0
packaging==24.2
pandas==2.3.3
patsy==1.0.2
pillow==12.0.0
pip==24.3.1
platformdirs==4.5.0
plotly==6.5.0
prometheus_client==0.23.1
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.1.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
PyJWT==2.10.1
pymdown-extensions==10.17.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.5
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.29.0
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==4.0.0
snowflake-snowpark-python==1.42.0
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.44
st-annotated-text==4.0.2
st-theme==1.2.3
statsmodels==0.14.5
streamlit==1.51.0
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.8
streamlit_faker==0.0.4
streamlit-image-coordinates==0.4.0
streamlit-keyup==0.3.0
streamlit-notify==0.3.1
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
websockets==15.0.1
wheel==0.45.1
xlrd==2.0.2
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
